requirements.txt
django==4.0.4

django-environ==0.9.0
django_cleanup==6.0.0
psycopg2-binary==2.9.3
django-tailwind==3.6.0
django-browser-reload==1.11.0
django-encrypted-model-fields==0.6.5

pandas==2.0.3

celery==5.2.7
redis==5.0.0
django-cors-headers==4.2.0
django-celery-beat==2.5.0

